Quarta fase do curso Laravel 5.1 com Angular

+++ app/Http/Controllers/ProjectController.php
@@ -3,7 +3,7 @@
 namespace CodeProject\Http\Controllers;
 
 use CodeProject\Repositories\ProjectRepository;
-
+use LucaDegasperi\OAuth2Server\Facades\Authorizer;
 use CodeProject\Services\ProjectService;
 use Illuminate\Http\Request;
 
@@ -26,18 +26,21 @@ class ProjectController extends Controller
     private $service;
 
 
+
+
     public function __construct(ProjectRepository $repository,ProjectService $service)
     {
         $this->repository = $repository;
         $this->service = $service;
+        $this->middleware('oauth');
 
     }
 
-
     public function index()
     {
+
 //     Restrives data function 'index' in 'ProjectService'  (Recupera os dados da função index)
-        return $this->service->index();
+        return $this->repository->findWhere(['owner_id'=>\Authorizer::getResourceOwnerId()]);
     }
 
     /**
@@ -48,6 +51,7 @@ class ProjectController extends Controller
      */
     public function store(Request $request)
     {
+
  //      Receives data and save the database (Recebe os dados e salva no banco de dados)
         //nota: Get all values this method '$request->all()' and send from class 'ProjectService' function 'create'
         //nota: Aqui pegamos todos os valores com o metodo '$request->all()' e enviamos para o a classe 'ProjectService' na função 'create'.
@@ -62,8 +66,13 @@ class ProjectController extends Controller
      */
     public function show($id)
     {
+
+        if($this->checkProjectPermissions($id)==false){
+           return ['error'=>'Access forbidden'];
+       }else{
 //      Returns the value of the id (Retorna o valor do id)
         return $this->service->show($id);
+       }
     }
   
     /**
@@ -75,6 +84,10 @@ class ProjectController extends Controller
      */
     public function update(Request $request, $id)
     {
+      
+        if($this->checkProjectPermissions($id)==false){
+            return ['error'=>'Access forbidden'];
+        }
 //        Retrieves the data of the project to according to ID and save the changes(Recupera os dados do project de acordo com ID e salva as alterações feitas)
         return $this->service->update($request->all(), $id);
     }
@@ -87,7 +100,9 @@ class ProjectController extends Controller
      */
     public function destroy($id)
     {
-       
+        if($this->checkProjectPermissions($id)==false){
+            return ['error'=>'Access forbidden'];
+        }
 //      Retrieves the data of the project and delete (Recupera os dados e os deleta)
         return $this->service->destroy($id);
     }
@@ -96,27 +111,62 @@ class ProjectController extends Controller
     
     public function showmembers($id)
     {
-
+        if($this->checkProjectPermissions($id)==false){
+            return ['error'=>'Access forbidden'];
+        }
         return $this->service->showmembers($id);
     }
     public function isMember($id, $memberId)
     {
+
+        if($this->checkProjectPermissions($id)==false){
+            return ['error'=>'Access forbidden'];
+        }
 
         return $this->service->isMember($id, $memberId);
     }
 
     public function storemembers(Request $request)
     {
+
+        if($this->checkProjectPermissions($request->project_id)==false){
+            return ['error'=>'Access forbidden'];
+        }
         return  $this->service->addMember($request->all());
     }
 
 
     public function destroymembers($id, $memberId)
     {
-       
-
+        
+        if($this->checkProjectPermissions($id)==false){
+            return ['error'=>'Access forbidden'];
+        }
         return $this->service->destroymembers($id, $memberId);
 
+    }
+//////////////////////// Initiate access validation (Inicia validação de acesso) ////////////////////////////////////
+    private function checkProjectOwner($projectID){
+
+        $userId = \Authorizer::getResourceOwnerId();
+        return $this->repository->isOwner($projectID, $userId);
+
+
+    }
+     private function checkProjectMember($projectID){
+
+            $userId = \Authorizer::getResourceOwnerId();
+            return $this->repository->hasMember($projectID, $userId);
+        }
+
+
+    private  function checkProjectPermissions($projectID){
+
+       
+        if($this->checkProjectOwner($projectID)or $this->checkProjectMember($projectID)){
+            return true;
+        }
+        return false;
     }
 
 }
